encapsulate the initial home menu in a component
